requirements.txt
opencv-python
streamlit
imagehash
numpy
ultralytics
pandas
altair
pillow
